requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --no-emit-index-url
#
attrs==23.1.0
    # via -r requirements.in
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
idna==3.4
    # via requests
python-gitlab==3.14.0
    # via -r requirements.in
pyyaml==6.0
    # via -r requirements.in
requests==2.30.0
    # via
    #   python-gitlab
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via python-gitlab
termcolor==2.3.0
    # via -r requirements.in
urllib3==2.0.2
    # via requests
